# docs/swagger.yaml
definitions:
  auth.AccessTokenRequest:
    properties:
      refreshToken:
        example: badbunybabybebe
        type: string
    required:
    - refreshToken
    type: object
  auth.LoginRequest:
    properties:
      email:
        example: mail@mail.com
        type: string
      password:
        example: Password123
        type: string
    required:
    - email
    - password
    type: object
  comment.Comment:
    properties:
      created_at:
        example: "2021-02-24 20:19:39"
        type: string
      deleted_at:
        example: "null"
        type: string
      id:
        example: 1099
        type: integer
      message:
        example: caption
        type: string
      photo_id:
        type: integer
      updated_at:
        example: "2021-02-24 20:19:39"
        type: string
      user_id:
        type: integer
    type: object
  comment.NewComment:
    properties:
      message:
        example: message
        type: string
      photo_id:
        type: integer
      user_id:
        type: integer
    required:
    - message
    - photo_id
    type: object
  comment.PaginationResultComment:
    properties:
      current:
        type: integer
      data:
        items:
          $ref: '#/definitions/comment.Comment'
        type: array
      limit:
        type: integer
      nextCursor:
        type: integer
      numPages:
        type: integer
      prevCursor:
        type: integer
      total:
        type: integer
    type: object
  controllers.MessageResponse:
    properties:
      message:
        type: string
    type: object
...
  user.DataSecurityAuthenticated:
    properties:
      expirationAccessDateTime:
        example: "2023-02-02T21:03:53.196419-06:00"
        type: string
      expirationRefreshDateTime:
        example: "2023-02-03T06:53:53.196419-06:00"
        type: string
      jwtAccessToken:
        example: SomeAccessToken
        type: string
      jwtRefreshToken:
        example: SomeRefreshToken
        type: string
    type: object
  user.DataUserAuthenticated:
    properties:
      email:
        example: some@mail.com
        type: string
      id:
        example: 123
        type: integer
      role:
        $ref: '#/definitions/user.Role'
      role_id:
        example: admin
        type: string
      userName:
        example: UserName
        type: string
    type: object
...
  user.Role:
    properties:
      created_at:
        example: "2021-02-24 20:19:39"
        type: string
      deleted_at:
        example: "null"
        type: string
      id:
        type: integer
      name:
        type: string
      updated_at:
        example: "2021-02-24 20:19:39"
        type: string
    type: object
  user.SecurityAuthenticatedUser:
    properties:
      data:
        $ref: '#/definitions/user.DataUserAuthenticated'
      security:
        $ref: '#/definitions/user.DataSecurityAuthenticated'
    type: object
info:
  contact: {}
paths:
  /auth/access-token:
    post:
      description: Auth user by email and password
      parameters:
      - description: body data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/auth.AccessTokenRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user.SecurityAuthenticatedUser'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.MessageResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.MessageResponse'
      summary: GetAccessTokenByRefreshToken UserName
      tags:
      - auth
  /auth/login:
    post:
      description: Auth user by email and password
      parameters:
      - description: body data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/auth.LoginRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user.SecurityAuthenticatedUser'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.MessageResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.MessageResponse'
      summary: Login UserName
      tags:
      - auth
  /comment:
    get:
      description: Get all Comments on the system
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/comment.PaginationResultComment'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.MessageResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.MessageResponse'
      security:
      - ApiKeyAuth: []
      summary: Get all Comments
      tags:
      - comment
    post:
      consumes:
      - application/json
      description: Create new comment on the system
      parameters:
      - description: body data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/comment.NewComment'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/comment.Comment'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.MessageResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.MessageResponse'
      security:
      - ApiKeyAuth: []
      summary: Create New CommentName
      tags:
      - comment
  /comment/{comment_id}:
    get:
      description: Get Comments by ID on the system
      parameters:
      - description: id of comment
        in: path
        name: comment_id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.MessageResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.MessageResponse'
      security:
      - ApiKeyAuth: []
      summary: Get comments by ID
      tags:
      - comment
...
swagger: "2.0"